chore: tweaks
