+++ plugins/plugin-seo/src/node/index.ts
@@ -1,3 +1,3 @@
 export * from './options.js'
-export * from './plugin.js'
-export * from './typings/index.js'
+export * from './seoPlugin.js'
+export * from '../typings/index.js'
