+++ plugins/plugin-seo/src/node/seoPlugin.ts
@@ -1,13 +1,14 @@
 import type { Plugin, PluginFunction } from 'vuepress/core'
 import { colors } from 'vuepress/utils'
-import { generateDescription } from './description.js'
-import type { SeoOptions } from './options.js'
-import { appendSEO, generateRobotsTxt } from './seo.js'
-import type { ExtendPage } from './typings/index.js'
+import type { ExtendPage } from '../typings/index.js'
+import { appendSEO } from './appendSEO.js'
+import { generateDescription } from './generateDescription.js'
+import { generateRobotsTxt } from './generateRobotsTxt.js'
+import type { SeoPluginOptions } from './options.js'
 import { logger, PLUGIN_NAME } from './utils/index.js'
 
 export const seoPlugin =
-  (options: SeoOptions): PluginFunction =>
+  (options: SeoPluginOptions): PluginFunction =>
   (app) => {
     if (app.env.isDebug) logger.info('Options:', options)
 
@@ -31,6 +32,6 @@ export const seoPlugin =
         appendSEO(app, options)
       },
 
-      onGenerated: (app): Promise<void> => generateRobotsTxt(app.dir),
+      onGenerated: (app): Promise<void> => generateRobotsTxt(app),
     }
   }
