+++ plugins/plugin-seo/src/typings/frontmatter.ts
