+++ plugins/plugin-seo/src/typings/index.ts
