+++ plugins/plugin-seo/src/typings/schema.ts
